refactor sql with constants

+++ src/main/java/com/robosh/model/dao/interfaces/AddressDao.java
@@ -5,7 +5,7 @@ import com.robosh.model.entity.Address;
 public interface AddressDao extends Dao<Address> {
     boolean checkAddressExist(String street, String numberHouse);
 
-    long getAddressId(String street, String numberHouse);
+    int getAddressId(String street, String numberHouse);
 
     Address getAddressByStreetNumberHouse(String street, String numberHouse);
 }
